Refactor: Move SnapSync tests to mod.rs and clean up
