--- crates/stages/stages/src/stages/snap_sync/tests.rs
@@ -1,211 +1,0 @@
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_utils::TestStageDB;
-    use alloy_primitives::B256;
-    use reth_network_p2p::{
-        download::DownloadClient,
-        snap::client::SnapClient,
-        priority::Priority,
-    };
-    use reth_network_peers::{PeerId, WithPeerId};
-    use reth_stages_api::{ExecInput, StageId, Stage};
-    use reth_config::config::SnapSyncConfig;
-    use reth_provider::{DatabaseProviderFactory, test_utils::MockNodeTypesWithDB};
-    use crate::stages::SnapSyncStage;
-    use std::sync::Arc;
-    use std::future;
-
-    /// Type alias for the provider used in tests
-    type TestProvider = reth_provider::DatabaseProvider<reth_db::mdbx::RW, MockNodeTypesWithDB>;
-
-    /// Simple mock snap client for testing
-    #[derive(Debug, Clone)]
-    struct MockSnapClient;
-
-    impl DownloadClient for MockSnapClient {
-        fn report_bad_message(&self, _peer_id: PeerId) {}
-        fn num_connected_peers(&self) -> usize { 1 }
-    }
-
-    impl SnapClient for MockSnapClient {
-        type Output = future::Ready<reth_network_p2p::error::PeerRequestResult<reth_eth_wire_types::snap::AccountRangeMessage>>;
-        
-        fn get_account_range_with_priority(
-            &self,
-            _request: reth_eth_wire_types::snap::GetAccountRangeMessage,
-            _priority: Priority,
-        ) -> Self::Output {
-            future::ready(Ok(WithPeerId::new(
-                PeerId::random(),
-                reth_eth_wire_types::snap::AccountRangeMessage {
-                    request_id: 1,
-                    accounts: vec![],
-                    proof: vec![],
-                },
-            )))
-        }
-
-        // Stub implementations for other methods (not used in our tests)
-        fn get_storage_ranges(&self, _request: reth_eth_wire_types::snap::GetStorageRangesMessage) -> Self::Output {
-            future::ready(Ok(WithPeerId::new(
-                PeerId::random(),
-                reth_eth_wire_types::snap::AccountRangeMessage {
-                    request_id: 1,
-                    accounts: vec![],
-                    proof: vec![],
-                },
-            )))
-        }
-
-        fn get_storage_ranges_with_priority(&self, _request: reth_eth_wire_types::snap::GetStorageRangesMessage, _priority: Priority) -> Self::Output {
-            self.get_storage_ranges(_request)
-        }
-
-        fn get_byte_codes(&self, _request: reth_eth_wire_types::snap::GetByteCodesMessage) -> Self::Output {
-            future::ready(Ok(WithPeerId::new(
-                PeerId::random(),
-                reth_eth_wire_types::snap::AccountRangeMessage {
-                    request_id: 1,
-                    accounts: vec![],
-                    proof: vec![],
-                },
-            )))
-        }
-
-        fn get_byte_codes_with_priority(&self, _request: reth_eth_wire_types::snap::GetByteCodesMessage, _priority: Priority) -> Self::Output {
-            self.get_byte_codes(_request)
-        }
-
-        fn get_trie_nodes(&self, _request: reth_eth_wire_types::snap::GetTrieNodesMessage) -> Self::Output {
-            future::ready(Ok(WithPeerId::new(
-                PeerId::random(),
-                reth_eth_wire_types::snap::AccountRangeMessage {
-                    request_id: 1,
-                    accounts: vec![],
-                    proof: vec![],
-                },
-            )))
-        }
-
-        fn get_trie_nodes_with_priority(&self, _request: reth_eth_wire_types::snap::GetTrieNodesMessage, _priority: Priority) -> Self::Output {
-            self.get_trie_nodes(_request)
-        }
-    }
-
-    #[test]
-    fn test_snap_sync_stage_creation() {
-        let config = SnapSyncConfig::default();
-        let snap_client = Arc::new(MockSnapClient);
-        let stage = SnapSyncStage::new(config, snap_client);
-        assert_eq!(<SnapSyncStage<MockSnapClient> as Stage<TestProvider>>::id(&stage), StageId::SnapSync);
-    }
-
-    #[test]
-    fn test_snap_sync_stage_disabled() {
-        let mut config = SnapSyncConfig::default();
-        config.enabled = false;
-        let snap_client = Arc::new(MockSnapClient);
-        let mut stage = SnapSyncStage::new(config, snap_client);
-        
-        let db = TestStageDB::default();
-        let provider = db.factory.database_provider_rw().unwrap();
-        let input = ExecInput { target: Some(100), checkpoint: None };
-        
-        let result = stage.execute(&provider, input);
-        assert!(result.is_ok());
-        assert!(result.unwrap().done);
-    }
-
-    #[test]
-    fn test_hashed_state_empty() {
-        let config = SnapSyncConfig::default();
-        let snap_client = Arc::new(MockSnapClient);
-        let stage = SnapSyncStage::new(config, snap_client);
-        
-        let db = TestStageDB::default();
-        let provider = db.factory.database_provider_rw().unwrap();
-        
-        let is_empty = stage.is_hashed_state_empty(&provider).unwrap();
-        assert!(is_empty);
-    }
-
-    #[test]
-    fn test_with_header_receiver() {
-        let config = SnapSyncConfig::default();
-        let snap_client = Arc::new(MockSnapClient);
-        
-        // Create a mock header with a specific state root
-        let mock_header = reth_primitives_traits::SealedHeader::new(
-            reth_primitives_traits::Header {
-                state_root: B256::from([0; 32]),
-                ..Default::default()
-            },
-            B256::ZERO
-        );
-        
-        let (sender, receiver) = tokio::sync::watch::channel(mock_header);
-        
-        let stage = SnapSyncStage::new(config, snap_client)
-            .with_header_receiver(receiver);
-        
-        // Test that header receiver is set
-        assert!(stage.header_receiver.is_some());
-        
-        // Test that we can get target state root
-        assert_eq!(stage.get_target_state_root(), Some(B256::from([0; 32])));
-    }
-
-    #[test]
-    fn test_create_account_range_request() {
-        let config = SnapSyncConfig::default();
-        let snap_client = Arc::new(MockSnapClient);
-        let mut stage = SnapSyncStage::new(config, snap_client);
-        
-        let request = stage.create_account_range_request(B256::ZERO, B256::from([0xff; 32]));
-        
-        assert_eq!(request.request_id, 1);
-        assert_eq!(request.starting_hash, B256::ZERO);
-        assert_eq!(request.limit_hash, B256::from([0xff; 32]));
-        assert_eq!(request.response_bytes, config.max_response_bytes);
-    }
-
-    #[test]
-    fn test_process_empty_account_ranges() {
-        let config = SnapSyncConfig::default();
-        let snap_client = Arc::new(MockSnapClient);
-        let stage = SnapSyncStage::new(config, snap_client);
-        
-        let db = TestStageDB::default();
-        let provider = db.factory.database_provider_rw().unwrap();
-        
-        let processed = stage.process_account_ranges(vec![]).unwrap();
-        assert_eq!(processed, 0);
-    }
-
-    #[test]
-    fn test_account_range_proof_verification() {
-        let config = SnapSyncConfig::default();
-        let snap_client = Arc::new(MockSnapClient);
-        let stage = SnapSyncStage::new(config, snap_client);
-        
-        // Test empty account range
-        let empty_range = reth_eth_wire_types::snap::AccountRangeMessage {
-            request_id: 1,
-            accounts: vec![],
-            proof: vec![],
-        };
-        assert!(stage.verify_account_range_proof(&empty_range).unwrap());
-        
-        // Test account range with accounts but no proof (should still pass with warning)
-        let range_with_accounts = reth_eth_wire_types::snap::AccountRangeMessage {
-            request_id: 1,
-            accounts: vec![reth_eth_wire_types::snap::AccountData {
-                hash: B256::from([1; 32]),
-                body: alloy_primitives::Bytes::new(),
-            }],
-            proof: vec![],
-        };
-        assert!(stage.verify_account_range_proof(&range_with_accounts).unwrap());
-    }
-}
