+++ llm/INDEX.md
@@ -1,0 +1,94 @@
+# SnapSync Implementation Documentation
+
+## 📚 Documentation Index
+
+This folder contains all LLM-generated documentation for the SnapSync stage implementation.
+
+### Core Documents
+
+1. **[FINAL_REVIEW.md](./FINAL_REVIEW.md)** - ⭐ **START HERE**
+   - Complete summary of work completed
+   - Test results and validation
+   - Code quality metrics
+   - Consistency analysis with other stages
+   - Production readiness assessment
+
+2. **[PROGRESS.md](./PROGRESS.md)**
+   - Detailed progress tracking during implementation
+   - Phase-by-phase completion status
+   - Issues found and resolved
+   - Current status and remaining work
+
+3. **[SNAP_SYNC_FINAL_SUMMARY.md](./SNAP_SYNC_FINAL_SUMMARY.md)**
+   - Earlier comprehensive summary
+   - Architecture overview
+   - Feature list
+   - Implementation details
+
+4. **[README.md](./README.md)**
+   - Quick overview
+   - Links to main documentation
+
+## 🎯 Quick Navigation
+
+### For Code Reviewers
+→ Start with [FINAL_REVIEW.md](./FINAL_REVIEW.md) for complete overview
+
+### For Understanding Implementation
+→ Read [SNAP_SYNC_FINAL_SUMMARY.md](./SNAP_SYNC_FINAL_SUMMARY.md) for architecture
+
+### For Tracking Progress
+→ Check [PROGRESS.md](./PROGRESS.md) for phase-by-phase work
+
+## ✅ Current Status
+
+**All work complete and verified:**
+- ✅ Tests restructured following reth patterns
+- ✅ Consistency verified against other stages  
+- ✅ All 6 unit tests passing
+- ✅ Clean compilation
+- ✅ Code follows reth standards
+
+**Test Results:**
+```
+running 6 tests
+test stages::tests::snap_sync_tests::test_create_account_range_request ... ok
+test stages::tests::snap_sync_tests::test_snap_sync_stage_basic_functionality ... ok
+test stages::tests::snap_sync_tests::test_process_account_ranges ... ok
+test stages::tests::snap_sync_tests::test_snap_sync_stage_creation ... ok
+test stages::tests::snap_sync_tests::test_snap_sync_stage_with_header_receiver ... ok
+test stages::tests::snap_sync_tests::test_snap_sync_stage_disabled ... ok
+
+test result: ok. 6 passed; 0 failed; 0 ignored; 0 measured
+```
+
+## 📊 Code Organization
+
+```
+/workspace/crates/stages/stages/
+├── src/
+│   ├── stages/
+│   │   ├── snap_sync.rs          # Main implementation
+│   │   └── mod.rs                # Tests (snap_sync_tests module)
+│   └── sets.rs                   # Pipeline integration
+└── Cargo.toml
+
+/workspace/crates/config/
+└── src/
+    └── config.rs                 # SnapSyncConfig definition
+```
+
+## 🔍 Key Changes Made
+
+1. **Test Location**: Moved tests from separate file to `mod.rs`
+2. **Public Fields**: Made config, header_receiver, request_id_counter public
+3. **MockSnapClient**: Implemented all 7 required trait methods
+4. **SnapSyncConfig**: Found and properly integrated existing config
+5. **Test Fixes**: All tests now passing with proper config setup
+
+## 📝 Notes
+
+- Documentation kept minimal and focused
+- All changes follow reth patterns
+- No redundant code or documentation
+- Ready for review and integration
