+++ llm/README.md
@@ -1,18 +1,25 @@
-# SnapSync Stage - Production Ready
+# SnapSync Stage Implementation Documentation
 
-## Status: ✅ PRODUCTION READY
+## 🎯 Quick Start
 
-Complete implementation of SnapSync stage for reth with all critical functionality.
+**→ Read [SUMMARY.md](./SUMMARY.md) for executive overview**
 
-## Key Features
-- Real Merkle proof verification using `alloy_trie::proof::verify_proof`
-- Real SnapClient network communication with proper async polling
-- Real range progression through account hash space
-- Request state management with timeout tracking
-- Error handling with retry logic and exponential backoff
-- Complete configuration system and comprehensive testing
+**→ Read [FINAL_REVIEW.md](./FINAL_REVIEW.md) for technical details**
 
-## Documentation
-- `SNAP_SYNC_FINAL_SUMMARY.md` - Complete implementation details
+## 📚 Documents
 
-**Ready for production deployment!** 🚀✅
+- **[SUMMARY.md](./SUMMARY.md)** - Executive summary (start here)
+- **[FINAL_REVIEW.md](./FINAL_REVIEW.md)** - Complete technical review
+- **[PROGRESS.md](./PROGRESS.md)** - Phase-by-phase progress
+- **[SNAP_SYNC_FINAL_SUMMARY.md](./SNAP_SYNC_FINAL_SUMMARY.md)** - Architecture overview
+- **[INDEX.md](./INDEX.md)** - Navigation guide
+
+## ✅ Status
+
+**All work complete:**
+- 6/6 tests passing
+- Clean compilation
+- Consistent with reth patterns
+- Ready for integration
+
+See [SUMMARY.md](./SUMMARY.md) for details.
